requirements.txt
gradio>=4.27.0
openai>=1.25.0
anthropic>=0.25.0
python-dotenv>=1.0.1
